#### 컨텐츠 로직 모두 구현완료. DONE:   - 모든 컨텐츠 구현완료   - 프레임 로직 구현완료   - 서버구조 구현완료 TODO:  - 테스트 해봐야함.     - 클라이언트 메시지들이 잘 해석되는지     - 세션들이 잘 지워지는지     - 플레이어들이 잘 삭제가 되는지     - 움직임이 틀어지지 않는지     - 공격판정이 제대로 되는지  - 오브젝트 풀 만들기 (링버퍼를 재활용하기 위해)

+++ NetLib/Player.cpp
@@ -22,26 +22,26 @@ void Player::OnPlayerDeath()
 	/*만약에 죽는효과가 있다면 여기다 구현*/
 }
 
-void Player::Move(const int x, const int y)
+void Player::Move(const short x, const short y)
 {
 	_X += x;
-	if (_X <= static_cast<int>(MAX_MAP_BOUNDARY::LEFT))
+	if (_X <= static_cast<short>(MAX_MAP_BOUNDARY::LEFT))
 	{
-		_X = static_cast<int>(MAX_MAP_BOUNDARY::LEFT);
+		_X = static_cast<short>(MAX_MAP_BOUNDARY::LEFT);
 		return;
 	}
-	if (_X >= static_cast<int>(MAX_MAP_BOUNDARY::RIGHT))
+	if (_X >= static_cast<short>(MAX_MAP_BOUNDARY::RIGHT))
 	{
-		_X = static_cast<int>(MAX_MAP_BOUNDARY::RIGHT);
+		_X = static_cast<short>(MAX_MAP_BOUNDARY::RIGHT);
 		return;
 	}
 	_Y += y;
-	if (_Y <= static_cast<int>(MAX_MAP_BOUNDARY::TOP))
+	if (_Y <= static_cast<short>(MAX_MAP_BOUNDARY::TOP))
 	{
 		_Y = static_cast<int>(MAX_MAP_BOUNDARY::TOP);
 		return;
 	}
-	if (_Y >= static_cast<int>(MAX_MAP_BOUNDARY::BOTTOM))
+	if (_Y >= static_cast<short>(MAX_MAP_BOUNDARY::BOTTOM))
 	{
 		_Y = static_cast<int>(MAX_MAP_BOUNDARY::BOTTOM);
 		return;
